linkedin_scraper/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic==2.8.2
pydantic-core==2.20.1

# Scraper dependencies
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# Fix for Python 3.12+ compatibility
setuptools>=65.0.0

# Optional: For enhanced logging and monitoring
structlog>=23.0.0

# AWS Lambda dependencies
mangum>=0.17.0
boto3>=1.34.0

# ATS Resume Parser dependencies
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0

# AI Analysis
anthropic>=0.40.0
openai>=1.0.0

# Auto-Apply: Job matching
rapidfuzz>=3.0.0
